proyecto4/requirements.txt
torch
transformers
peft
bitsandbytes
accelerate
scipy
trl
datasets
sentencepiece
protobuf
pandas
numpy
huggingface_hub
python-dotenv
